pyproject.toml
[project]
name = "tool-usage-evals"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "azure-identity>=1.23.0",
    "openai>=1.82.0",
    "pydantic>=2.11.5",
    "python-dotenv>=1.1.0",
]

[build-system]
requires = ["hatchling", "uv-dynamic-versioning"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/tool_usage_evals"]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
]
